+++ cmd/ichsm/main.go
@@ -66,7 +66,7 @@ func newRootCommand(out io.Writer, errOut io.Writer) *cobra.Command {
 	cmd.SetOut(out)
 	cmd.SetErr(errOut)
 	cmd.SetVersionTemplate("{{.Name}} {{.Version}}\n")
-	cmd.AddCommand(newSearchCommand(), newQueryCommand(), newMatchCommand(), newSummaryCommand(), newReadsCommand(), newLinksCommand(), newPubsCommand(), newOpenCommand(), newIdentifyCommand(), newGetFieldsCommand(), newGetValuesCommand())
+	cmd.AddCommand(newSearchCommand(), newQueryCommand(), newMatchCommand(), newSummaryCommand(), newReadsCommand(), newLinksCommand(), newPubsCommand(), newOpenCommand(), newIdentifyCommand(), newGetFieldsCommand(), newGetValuesCommand(), newUpdateCommand())
 	return cmd
 }